% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baixa-docs.R
\name{download_documents}
\alias{download_documents}
\title{Download PDF documents belonging to lawsuits}
\usage{
download_documents(id, path, login = NULL, password = NULL,
  only_petitions = FALSE, verbose = FALSE)
}
\arguments{
\item{id}{A character vector with one or more lawsuit IDs}

\item{path}{Path to directory where to save files}

\item{login}{ESAJ system login (if left `NULL`, will ask for it)}

\item{password}{ESAJ system password (if left `NULL`, will ask for it)}

\item{only_petitions}{Whether to only download petitions}

\item{verbose}{Whether to output messages while downloading}
}
\description{
Download PDF documents belonging to lawsuits
}
